Move classes to templates module
