+++ qubesadmin/templates.py
@@ -20,6 +20,8 @@
 """Template management for Qubes OS."""
 
 import os
+import datetime
+import enum
 import fnmatch
 import subprocess
 import tempfile
@@ -28,9 +30,72 @@ import sys
 
 import qubesadmin.exc
 
+DATE_FMT = '%Y-%m-%d %H:%M:%S'
 PATH_PREFIX = '/var/lib/qubes/vm-templates'
 PACKAGE_NAME_PREFIX = 'qubes-template-'
 TAR_HEADER_BYTES = 512
+
+
+class TemplateState(enum.Enum):
+    """Enum representing the state of a template."""
+    INSTALLED = 'installed'
+    AVAILABLE = 'available'
+    EXTRA = 'extra'
+    UPGRADABLE = 'upgradable'
+
+    def title(self) -> str:
+        """Return a long description of the state. Can be used as headings."""
+        # pylint: disable=invalid-name
+        TEMPLATE_TITLES = {
+            TemplateState.INSTALLED: 'Installed Templates',
+            TemplateState.AVAILABLE: 'Available Templates',
+            TemplateState.EXTRA: 'Extra Templates',
+            TemplateState.UPGRADABLE: 'Available Upgrades'
+        }
+        return TEMPLATE_TITLES[self]
+
+
+class VersionSelector(enum.Enum):
+    """Enum representing how the candidate template version is chosen."""
+    LATEST = enum.auto()
+    """Install latest version."""
+    REINSTALL = enum.auto()
+    """Reinstall current version."""
+    LATEST_LOWER = enum.auto()
+    """Downgrade to the highest version that is lower than the current one."""
+    LATEST_HIGHER = enum.auto()
+    """Upgrade to the highest version that is higher than the current one."""
+
+
+# pylint: disable=too-few-public-methods,inherit-non-class
+class Template(typing.NamedTuple):
+    """Details of a template."""
+    name: str
+    epoch: str
+    version: str
+    release: str
+    reponame: str
+    dlsize: int
+    buildtime: datetime.datetime
+    licence: str
+    url: str
+    summary: str
+    description: str
+
+    @property
+    def evr(self):
+        """Return a tuple of (EPOCH, VERSION, RELEASE)"""
+        return self.epoch, self.version, self.release
+
+
+class DlEntry(typing.NamedTuple):
+    """Information about a template to be downloaded."""
+    evr: typing.Tuple[str, str, str]
+    reponame: str
+    dlsize: int
+
+
+# pylint: enable=too-few-public-methods,inherit-non-class
 
 
 def qubes_release() -> str:
